apps/backend/requirements.txt
Django==4.2.9
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
django-filter==23.5

# Database
psycopg2-binary==2.9.9

# Redis & Celery
redis==5.0.1
celery==5.3.6
django-celery-beat==2.6.0

# Image handling
Pillow==10.2.0

# Utilities
python-dotenv==1.0.0
nanoid==2.0.0

# Production
gunicorn==21.2.0
whitenoise==6.6.0

# Fingerprinting
user-agents==2.2.0

# Phone validation
phonenumbers==8.13.27

# Storage (AWS S3)
boto3==1.34.14
django-storages==1.14.2

# Payment
razorpay==1.4.2

# SMS & HTTP
requests==2.31.0

# WebSocket Support (Django Channels)
channels>=4.0.0
channels-redis>=4.1.0
daphne>=4.0.0

# Testing
pytest==7.4.4
pytest-django==4.8.0
pytest-asyncio>=0.21.0
factory-boy==3.3.0

# Load Testing
locust==2.20.1

# AI Services
openai>=1.0.0
google-cloud-vision>=3.4.0
Pillow>=10.0.0

# ML/Data Processing
scikit-learn>=1.3.0
numpy>=1.24.0

# Image Processing
opencv-python>=4.8.0

# Color Analysis
colour>=0.1.5

# ASGI Server
uvicorn[standard]>=0.24.0
